Moved accounts page to its own file
